+++ src/application/use-cases/LogIn.ts
@@ -1,15 +1,14 @@
 import {AuthRepository} from "../../domain/repositories/AuthRepository.ts";
-import {User} from "../../domain/entities/User.ts";
 
 export class LogIn {
     constructor(private authRepository: AuthRepository) {
     }
 
-    async execute(username: string, password: string): Promise<User> {
-        const user = await this.authRepository.logIn(username, password);
-        if(user) {
-            this.authRepository.saveToken(user);
+    async execute(username: string, password: string): Promise<boolean> {
+        const token = await this.authRepository.logIn(username, password);
+        if(token) {
+            this.authRepository.saveToken(token);
         }
-        return user;
+        return true;
     }
 }
